CI4: Bugfix correct the removal of env backup
The previous version of the migration script only removed the backup under one set of circumstances. This version attempts to  remove the backup in all cases once the script either succeeds and returns from function up(), or fails and redirects back to login from function convert_ci3_encrypted_data(). We now have to test if the file exists before attempting to remove it in the security_helper function remove_backup.

+++ app/Helpers/security_helper.php
@@ -94,6 +94,10 @@ function abort_encryption_conversion()
 function remove_backup()
 {
 	$backup_path = WRITEPATH . '/backup/.env.bak';
+	if( ! file_exists($backup_path))
+	{
+		return;
+	}
 	if(unlink($backup_path) === false)
 	{
 		log_message('error', "Unable to remove $backup_path.");
